move install and benchmark in to folders
