+++ sotopia/cli/benchmark/__init__.py
@@ -1,0 +1,3 @@
+from .benchmark import benchmark
+
+__all__ = ["benchmark"]
