+++ sotopia/cli/install/__init__.py
@@ -1,0 +1,3 @@
+from .install import install
+
+__all__ = ["install"]
